refactor: extract planStart into handlers module
